Fix: Fake timers were swallowing exceptions
Currently clock.tick() was executing fake timer callbacks inside a
try...catch block which would silently swallow any exceptions which
may occur in the callback.

This patch changes this behavior to re-throw any exception that might
occur. If there are multiple timer callbacks to be executed, and there
are multiple exceptions, only the first one will be thrown, but all
timers will be executed.

Fixes #27