Allow yields* and callsArg* to create sequences of calls.
Calling yields, callsArg, or their variants multiple times will now create a sequence of actions, executed one after the other. The last action will be repeated. For example:

var stub = sinon.stub().yields(1).callsArgWith(0, 2).yields(3);

will result in a stub that, on first invocation, yields 1; on second invocation, calls its zeroth argument with 2; and on the third invocation and all others after it, yields 3. Such things can be useful, for example, when testing behavior like retry-for-up-to-three-failures.

This is backward compatible, since the last action is repeated, so if only one action is specified then the stub will behave just as before.

Closes #157.