Adds logging to the FakeServer
To enable (in node or a modern browser):
server.enableResponseLogging = true;
sinon.log = console.log;

This adds to the fake server the capability of logging responded
requests. This feature is opt in. What will be printed is a beautified
stringify of the request and response objects.

If the current environment does not support JSON.stringify the logging
fails silently.