Adde onLastCall function
When tests change the behavior of a stub, we have to mention the index of the call to the stub.

Most of the time a test is only concerned with the latest behavior, which is onCall(this.behaviors.length)

I have added this convenience function along with the existing onFirstCall, onSecondCall, onThirdCall functions.