Fix #777: Support non-enumerable props when stubbing objects
This commit modifies sinon.stub() such that when only an object is
given, all non-enumerable properties as well as enumerable properties
that are functions will be stubbed.

The iteration over the object is accomplished by using sinon.walk, a new
utility function that supports iterating over all properties of an
object up the prototype chain, falling back to for..in when
getOwnPropertyNames is not available.

This functionality allows sinon.stub() - and especially methods like
sinon.createStubInstance() - to be future-forward by facilitating the
fact that methods defined in class declarations are non-enumerable by
default. It also allows stubbing of native objects such as Arrays and
Dates without having to explictly pass property names.