Prevent false negatives in walk-test due to for..in behavior
Different environments are inconsistent in how they handle for..in. To
compensate for this the test in which covers the case where walk uses
for..in in enumerates through the target object manually, keeping track
of the number of expected calls so that the assertion at the end holds
regardless of environment.