lib/sinon/util: Remove window conditionals from IE files.
The removed conditionals should not have any functional benefit, so in
that sense it's safe to remove these. (As described in #840.)

However, it appears that these conditionals have an unintended effect
in IE 11 and Edge 12, where the functions defined within the
conditionals actually stop working as expected. (See #848.)

While it is true these IE utils are not necessary for the newer IEs,
they are often still loaded as part of a generic test package.

Additionally, to make the ci-tests work:

- Renamed util/*_ie.js to util-ie/*.js.
- Made sure IE utils are not included in node tests.
  They were included before, but didn't actually run because of the
  `if (typeof window !== 'undefined')` conditional. With the conditional
  gone, just don't include them.

Closes #840.
Fixes #848.