Ensure sinon can run in a  WebWorker
Related to the fixes in #908.

It turned out that although we support running Sinon inside
a WebWorker, we do not actually test this in any of the tests.
This test will be run by the test config of the packaged build.

As PhantomJS 1.9.8 does not implement WorkerGlobalScope, the
test if we are running in a WebWorker is a bit simplified, but
it works for us.