Extract `util/fake_xml_http_request` into a CommonJS module.
This module no-longer mutates the `sinon` object; instead it exports an API which is then added to the `sinon` global by the `lib/sinon.js` entrypoint.

This module still relies on the sinon object as the tests are tightly coupled to `sinon.logError` (they stub it to prevent uncaught exceptions); we should be able to resolve this once the tests are CommonJSified.

I have reservations on the way `useFakeXMLHttpnRequest` mutates `FakeXMLHttpRequest` however I feel this is best refactored post the CommonJS transition.