Extract `util/fake_server` into a CommonJS module.
This module no-longer mutates the `sinon` object; instead it exports an API which is then added to the `sinon` global by the `lib/sinon.js` entrypoint.

This module still depends on the `sinon` object as the tests are tightly coupled to `sinon.useFakeXMLHttpRequest` and friends; once the tests are migrated to CommonJS we should be able to rectify this.