Migrate Packaged Tests to use a Browserified Build.
Browserify-shim has been added to take the global buster instance provided by the test-runner (rather than including it in the browserified bundle).  Tests are bundled into `tmp/` ahead of being executed.

`issues-test.js` had to change to have a stricter `requiresSupportFor` check as `process` is provided by browserify

`fake-xml-http-request-test.js` had to change to pick up the XMLHttpRequest instance from the global (window) rather than trying to get it from the `this` value in the closure.

Chagned the `'use strict'` requirement in the `test/` folder to be of the global form (typical of CommonJS modules) and moved all web-worker tests into their own folder as they require the `function` form of `'use strict'` and a built version of the code (at least for now...)