Convert values to strings with toString instead of String()
On Node v0.12 the convertion from Symbol to string using the explicit
cast with String throws a TypeError. To perform this conversion the
toString method on the Symbol must be used.

The toString method can on the other hand not be invoked on null or undefined
values. For these we can use the explicit cast to String.

This commit adds an auxilary method for converting all values to string and uses
to convert user supplied values before string concatenation.