Add missing error event listener key to the FakeXMLHttpRequest base class.
Without this, the changes in commit 2cfbacd5cea5b63c014076d3a65b6642b2200793 cause
an onerror function handler on the FakeXMLHttpRequest instance to never be called,
since the "error" event is triggered but there is no event listener listening for
it. This results in regular error handlers not being called for legitimate error
responses (like 500 Internal Server Error).