ReactNative compatibility. Allow sinon fakeServer to run in React Native (#1052)
ReactNative compatibility. Allow sign fakeServer to run in React Native environment either on device or browser debug.

ReactNative has xhr support via polyfill. But the check for supporting xhr looked and withEntitlements which is failing in RN.

ReactNative does not have window.location so a default assignment and check for undefined window.location was needed.