Do not invoke getters in walk (#1059)
When stubbing a prototype, the `walk` implementation accesses getters to
obtain the value of each property to pass it to the iterator function.
This value is not needed by any of the iterator implementations anymore,
but invoking the getter may have side effects or throw exceptions.

With this change, values are no longer obtained from the objects. It
keeps the code to pass a different target object to the iterator which
allows to obtain the value in the iterator implementation if necessary.