Fix a regression introduced in #692
Spying on the methods of the localStorage object failed to
work in Chrome when Sinon started using `Object.defineProperty`.
Instead, the effect was to add a key with the name of the method
(say "getItem") to the disk storage, instead of on the actual
localStorage object.

This seems to only affect Chrome.

Closes #1046