Revert #1017 fix-fake-xhr-compliance-again"
This reverts commit e35c9d5244fa853b62f6f2464592d61ba3b27f66,
reversing changes made to 568a4be09befcf3cc582010205831e9e1806cf85.

The changes in the networking api proved to be too big of a change
for version 1.17, and was decided to be kept only for the upcoming
version 2 in the master branch.

See the discussion in #1040 for more background info:
https://github.com/sinonjs/sinon/issues/1040#issuecomment-231727271