Add .resolves and .rejects to stub
This commit adds the possibility for a stub
to be configured to return a promise which will
either resolves or reject

This was already possible, but this adds a shorthand for it

stub.returns(Promise.resolve({})) -> stub.resolve({})
stub.returns(Promise.reject(new Error("bad luck"))) -> stub.rejects("bad luck")

Since Promise are now part of the standard ES6 features, this
commit does not make use of polyfill or third party library.
This is problematic for some browsers and generally for contexts where
the standard is not implemented.