Abort pre-commit script when no files are changed
Running this script when there are no files to stash
will mess up the current files and pop previous
stashes from the stash list.

Exit early to avoid the mess: it makes little sense to do a
linting of changed files when there are none, so it should
be seen as a mistake anyway.