Remove polyfill for Promise
The ramifications of letting the Promise polyfill enter the main
Sinon code escaped me at the time I reviewed #1211, but it struck me
when reviewing #1215 that had similar issues. In short:

    having the polyfill bundled with Sinon will lead to false
    negatives: running tests using Sinon will then give
    the impression that code works on platforms that does not
    support the features that were polyfilled.

This reverts the introduction of `native-promise-only` in the main
code and moves it to its proper place: the test code.