Split spy-formats into a separate package
This change splits two concerns into separate files, which will make it a little
easier to discern which one changes the more often, and which one is more
bug-prone. The change was made as a response to bithound complaining about 1)
long file, and 2) bug-prone file (high percentage of commits are correcting
bugs).