Fix issue #1332: little bug correction in spy.printf "%*" formatter.
When used with a list of strings, only the first one is correctly non-quoted :
spy.printf("%*", "a", "b", "c") -> a, "b", "c"

The formatio.ascii function was called with all the arguments provided by map (value, index, collection).
Just added a wrapper in the map() call to forward only the first argument (value).

The unit test shows that the error is a little more serious:
in case the arguments are not simple values, but array/object, formatio throws an exception.