Add proxyquire-universal to headless run
This commit reintroduces the proxyquire-universal changes that
caused some issues (see #1498 and commit cbf11282).

What is different is that it removes proxyquire from the build step.
This should never have been there, and was probably introduced to
make proxyquire calls in the tests work with browsers. The problem
was that the tests are not using the built sinon file, so the
config changes would only be utilized for the built file.

By adding the proxquire-universal plugin to the NPM script "test-headless"
it would only take place for these tests and the cloud tests (which are
reusing the headless config).

This commit, along with the previous, fixes #1498.