Simplify scripts using lifecycle events
The WebWorker test in PR #1479 seems to fail for some mysterious
reason with a synchronous XHR requests. The reason was most
probably the importScript('../../pkg/sinon.js') call in the
webworker file was referencing a build file that wasn't built.

By now explicitly building this file as part of the test run
this error has now disappeared locally, and probably will
disappear on CircleCI as well.

fixes the build in #1479