Add props for default file in CDN delivery
Some NPM based CDNs look at the package.json to find
out which file to deliver as the default file to
browsers. See for instance
https://github.com/jsdelivr/jsdelivr#configuring-a-default-file-in-packagejson

Using the 'browser' field would be wrong, as it does _not_
signify a bundle, only an entry point for bundlers.

This will basically only work for a few CDNs, but does not
cost us anything in maintenance.