Docs: verify documentation using markdownlint
This commit bundles a custom version of markdownlint-cli, which includes
markdownlint@0.6.1.

This can be replaced by a regular release of markdownlint-client, once
https://github.com/igorshubovych/markdownlint-cli/pull/21 is merged and
a version is published.

markdownlint is now used for verifying all the markdown files in `docs/`
as a precommit hook and a task that can be used by CI.