Revert spied fakeTimers to original state
When sandbox.useFakeTimers() is called, the original setTimeout and
other methods are replaced with fake methods. When such fake method is
spied it's replaced again with another fake. To make sure that
sandbox.restore() reverts setTimeout and other methods to the original
state these fakes should be restored in reverse order to make sure each
fake is restored to its predecessor when calling fake.restore().