Adds guard for empty properties in deepEqual when a matcher is provided
In JavaScript every acts like the "for all" quantifier in mathematics.
In particular, for an empty array, it returns true. (It is vacuously
true that all elements of the empty set satisfy any given condition.)

This is not the intended behaviour for deepEqual, we should treat empty
properties as falsy.

Resolves #1900