Remove misleading 'own' from exception message
The property descriptor is recursively retrieved up the
prototype chain. It can by thus often _not_ be owned by
the object. Therefore the previous error message was wrong.
We even have an explicit test for the sandbox.replace* methods
that states this:
    it("should replace an inherited property", ...

This is in opposition to stubs and spies which refuse to replace
fields that are not owned by the object, which is somewhat strange ...