Remove npm-run-all
In 0d1e40bf6d1603791764330947c25e7877f4956c the
last use of `run-p` was removed, which is arguably
the whole justification for this package.

It is trivial to run commands *serially* using
`npm run` and `&&`, making this package
superfluous.

Removing it, reduces the maintenance burden
slightly and will speed up package install
slightly.