Use `util.inspect` for formatting
As part of a larger effort, we are retiring
`@sinonjs/formatio`. When it was created Node was
quite immature and there were no other community
projects for solid object formatting.

Ten years later, Node is mature and has great
support for formatting all sorts of values.

Removing the use of `@sinonjs/formatio` in favour
of Node's `util.inspect` removes an entire
project's worth of maintenance burden from the
organisation.

In practial terms, it means that some output from
`sinon` will look slightly different. This won't
cause a new major release as this output is only
intended to be read by humans and not be part of
an API promise.