[Docs] Update fake-server usage & descriptions
With the update to support es6 modules, the organization of `nise` module inside of sinon namespace changed.  `createFakeServer()` no longer exists but rather `fakeServer` is a property of an imported/required module `sinon` with a singular `create()` method.

This is updated on the main page of https://sinonjs.org but not in the detailed API page here.  

Added additional clarity and important information for usage of parameters.  Removed duplication of property descriptions to increase readability.