Re-fix the comment about appending stack frames
The previous "fix" was wrong, I misunderstood this code and the comment.
What's actually happening here is that we want to add a frame of context
to `callStr`, but the first two stack frames will be within Sinon code
and thus probably not helpful to the end-user.

So, we skip the first two stack frames, and append the third stack
frame, which should contain a meaningful location to the end-user.