Revert breaking change to stack-stripping regex
This regex was made more specific in an attempt to avoid incorrectly
stripping off parts of messages - we strip off anything after " at",
which means "to have been called at least once" becomes "to have been
called".

However, the new regex was too specific and made faulty assumptions
about stack traces - namely, that they'll always have parentheses after
the at. Firefox's stack traces, however, look like so:

func() at callFn@about:blank line 2 > injectedScript:47353:21

So, we could try to come up with a better regex that matches all
browsers that we test with, or perhaps add some sort of indicator when
we're appending the stack traces, but since this change is breaking the
tests for everyone right now, for expedience we'll just revert to how we
were stripping things before.

This means changing one of the asserts back to how it was before, where
it was asserting against a partial message due to the over-eager
stripping, so add a comment about that as well.