Move changelog updates into version script
Previously we were updating the changelog and doing other auxiliary
version bumps in the post-version script, which meant they weren't
included in the tagged release commit, and required a followup commit.

This moves the version-bumping changes into a new `version.sh` script,
so that those changes show up in the tagged version commit.