Update with a more general regex
This should match any stack frame format that has at least one non-word,
non-space character in it, which should be a pretty safe assumption, and
holds for the three browsers we test in anyway.

This gets us back to not trimming messages improperly, while also not
breaking firefox, and I think is reasonable as far as regexes go.

This seems like a reasonable enough regex - the original fixed one was
overly complicated because it tried to remove an arbitrary number of
stack frames, which was unnecessary because we only append one.