Adjust deploy scripts to archive old releases in a separate branch, move existing releases out of master (#2426)
* Remove release archives from master

These archives made it difficult to find things in the GitHub interface,
and take up a lot of space in a checked-out repo for something that is
not useful to most people checking out the repository.

The main purpose of these archives is to make old versions and
documentation available on the Sinon website that is run out of this
repo. This can be supported by using a separate branch for website
releases, and to maintain the archives.

Following this commit, the `npm version` scripts will be updated to
automatically handle archiving the releases in the new releases branch
and keeping it up to date with master.

Also remove the directories we removed from .prettierignore, since they
don't exist any more.

* Update npm version scripts to manage new releases branch

This updates the `postversion.sh` script to merge master into the new
`releases` branch and then copy and commit the version archives there,
instead of into master, so that the archives aren't cluttering up the
master branch unnecessarily.

* Move changelog updates into version script

Previously we were updating the changelog and doing other auxiliary
version bumps in the post-version script, which meant they weren't
included in the tagged release commit, and required a followup commit.

This moves the version-bumping changes into a new `version.sh` script,
so that those changes show up in the tagged version commit.

* Don't ignore changelogs in prettier

We were manually running prettier on CHANGELOG.md during release but
ignoring it in .prettierrc - this may have been a holdover from before
Prettier had the proseWrap option, which would have wreacked havoc with
the changelog, but the current version handles things just fine.

* Merge silently into the archive branch

We shouldn't ever have conflicts here, so we can just auto-merge

* Tweak changelog.md generation to be more compatible

$(<CHANGES.md) wasn't working on my machine, I think because it's a
bash-ism, and my machine was running it under sh. I also took the chance
to do something more useful than just `cat` by removing the double
header that was previously on this page.

* Fail if we cancel CHANGES.md

Also make version.sh a bash script like the others - this would have
fixed the update-change-log-page script, in retrospect, but I think it's
worth getting rid of the double header anyway.

* Clean up copied directories

Otherwise when we switch back to the source branch we'll have these
files left dangling, which was what we were trying to avoid in the first
place

* Don't defeat set -e

Stringing commands together like this makes it so if the push doesn't
work, we keep going, which we don't want!

* Temporary commit to make publish a dry-run

This makes the script work on my machine, since I of course can't
actually publish to sinon's npm package

Also skip postbuild because chrome tests aren't working on my machine
and `npm install` doesn't fix it

* 12.0.2

Co-authored-by: Carl-Erik Kopseng <carlerik@gmail.com>