breaking: Remove sinon.defaultConfig and related modules (#2565)
* breaking: Remove sinon.defaultConfig and related modules

default-config and get-config are leftovers from when Sinon
shipped with sinon.test (now the independent NPM module
 'sinon-test').

These serve no purpose internally, and really have no purpose
but to help sinon-test create a base default. If needed,
these can be copied into the sinon-test project. No projects
should depend on these (my assumption), but since it is a
change of the API we mark it as a breaking change

fixes #2561

* fixed up tests that were lying

It seemed like the the 'injectInto' option
would expose most props by default. This was
not the case. That was formerly hidden by
using the getConfig call that added props
that were never used in the actual implementation.

Added another test to make this more explicit.
Will add docs on this.

* Document Sandbox#inject

This was added in Sinon 0.6 but has never been documented

* Note that 'properties' is empty by default