fix #2589: avoid invoking getter as side-effect
of creating restorerer function. Check for the odd case
where we actually force setting a value through accessor
functions and only look up the value at that time.

The alternative would be more elaborate, check if the
descriptor had a `get` or a `value` field, and then
decide, but this should do the trick.