chore!: remove assert.failException property
This was used for allowing customisation of the thrown error between
sandboxes and for customisation for integrations (most likely from the
days of BusterJS).

To my knowledge and what I've been able to find on GitHub, this has
never had any _production_ use by end users.

BREAKING CHANGE: this removes assert.failException from the API